Add stuff for ridable entities, make keybindings contextual based on if you're riding an entity

+++ src/main/java/tragicneko/tragicmc/entity/mob/EntityJetNeko.java
@@ -61,7 +61,7 @@ public class EntityJetNeko extends EntityNeko {
 
 	@Override
 	public int getTotalArmorValue() {
-		return (int) 6;
+		return 6;
 	}
 
 	@Override
